+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/abc/ScriptPack.java
@@ -541,6 +541,7 @@ public class ScriptPack extends AS3ClassTreeItem {
             FileTextWriter writer2 = exportSettings.singleFile ? exportSettings.singleFileWriter : writer;
             ConvertData convertData = new ConvertData();
             convertData.ignoreFrameScripts = exportSettings.ignoreFrameScripts;
+            convertData.ignoreAccessibility = exportSettings.ignoreAccessibility;
             convertData.exportEmbed = exportSettings.exportEmbed;
             convertData.exportEmbedFlaMode = exportSettings.exportEmbedFlaMode;
             convertData.assetsDir = exportSettings.assetsDir;